Add file system reset and cache invalidation
File systems track ISleetFiles in memory which point to files in the local cache. If the same file system is reused between commands/locks it can become out of date and will end up using older versions of files. To avoid this the file system should be reset after each lock. Files taken from before the reset should be invalidated and should fail when used to avoid problems.