Add support for configuring with .netconfig
The [dotnetconfig](https://dotnetconfig.org) format allows hierarchical
tool settings to be persisted and read in a standards way by multiple
tools simultaneously. By adding this support to the tool, all settings
can now also be persisted to the`.netconfig` in the current directory
to streamline subsequent runs. In addition, for scenarios where multiple
projects/repositories in the local machine produce packages that are
pushed to the same feed(s), the `.netconfig` support makes it trivial
to reuse those settings across the entire machine by saving the settings
to the user profile instead.

Updated the docs to cover both `sleet.json` and `.netconfig` formats
equally.